// src/generated/models/Mserp_dirpersonuserentitiesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export interface Mserp_dirpersonuserentities {
  mserp_dirpersonuserentityid: string;
  mserp_partynumber: string;
  mserp_personname?: string;
  mserp_personnamealias?: string;
  mserp_personprimaryemail?: string;
  mserp_primaryfield?: string;
  mserp_useremail?: string;
  mserp_userid: string;
  mserp_validfrom?: string;
  mserp_validto?: string;
}