// src/generated/models/Mserp_hcmemployeev2entitiesModel.ts
export interface Mserp_hcmemployeev2entities extends Mserp_hcmemployeev2entitiesBase {
  mserp_allowrehirename?: string;
  mserp_employmentnoticeemployerunitname?: string;
  mserp_employmentnoticeworkerunitname?: string;
  mserp_fk_person_idname?: string;
  mserp_fk_worker_idname?: string;
  mserp_gendername?: string;
  mserp_isdisabledname?: string;
  mserp_isdisabledveteranname?: string;
  mserp_isexpatriaterulingapplicablename?: string;
  mserp_isfulltimestudentname?: string;
  mserp_maritalstatusname?: string;
  mserp_name?: string;
  mserp_partytype?: string;
  mserp_primarycontactemailisimname?: string;
  mserp_primarycontactemailisprivatename?: string;
  mserp_primarycontactfacebookisprivatename?: string;
  mserp_primarycontactfaxisprivatename?: string;
  mserp_primarycontactlinkedinisprivatename?: string;
  mserp_primarycontactphoneismobilename?: string;
  mserp_primarycontactphoneisprivatename?: string;
  mserp_primarycontacttwitterisprivatename?: string;
  mserp_primarycontacturlisprivatename?: string;
  mserp_worksfromhomename?: string;
  mserp_fk_person_id?: object;
  _mserp_fk_person_id_value?: string;
  mserp_fk_worker_id?: object;
  _mserp_fk_worker_id_value?: string;
}