// src/generated/models/Mserp_hcmworkerentitiesModel.ts
export const Mserp_hcmworkerentitiesmserp_allowrehire = {
  200000000: 'No',
  200000001: 'Yes'
} as const;
export type Mserp_hcmworkerentitiesmserp_allowrehire = keyof typeof Mserp_hcmworkerentitiesmserp_allowrehire;
export const Mserp_hcmworkerentitiesmserp_gender = {
  200000000: 'None',
  200000001: 'Male',
  200000002: 'Female',
  200000003: 'Non_Specific'
} as const;
export type Mserp_hcmworkerentitiesmserp_gender = keyof typeof Mserp_hcmworkerentitiesmserp_gender;
export const Mserp_hcmworkerentitiesmserp_isdisabled = {
  200000000: 'No',
  200000001: 'Yes'
} as const;
export type Mserp_hcmworkerentitiesmserp_isdisabled = keyof typeof Mserp_hcmworkerentitiesmserp_isdisabled;
export const Mserp_hcmworkerentitiesmserp_isdisabledveteran = {
  200000000: 'No',
  200000001: 'Yes'
} as const;
export type Mserp_hcmworkerentitiesmserp_isdisabledveteran = keyof typeof Mserp_hcmworkerentitiesmserp_isdisabledveteran;
export const Mserp_hcmworkerentitiesmserp_isexpatriaterulingapplicable = {
  200000000: 'No',
  200000001: 'Yes'
} as const;
export type Mserp_hcmworkerentitiesmserp_isexpatriaterulingapplicable = keyof typeof Mserp_hcmworkerentitiesmserp_isexpatriaterulingapplicable;
export const Mserp_hcmworkerentitiesmserp_isfulltimestudent = {
  200000000: 'No',
  200000001: 'Yes'
} as const;
export type Mserp_hcmworkerentitiesmserp_isfulltimestudent = keyof typeof Mserp_hcmworkerentitiesmserp_isfulltimestudent;
export const Mserp_hcmworkerentitiesmserp_maritalstatus = {
  200000000: 'None',
  200000001: 'Married',
  200000002: 'Single',
  200000003: 'Widowed',
  200000004: 'Divorced',
  200000005: 'Cohabiting',
  200000006: 'Registeredpartnership',
  200000007: 'Separated',
  200000008: 'Civilpartnership'
} as const;
export type Mserp_hcmworkerentitiesmserp_maritalstatus = keyof typeof Mserp_hcmworkerentitiesmserp_maritalstatus;
export const Mserp_hcmworkerentitiesmserp_primarycontactemailisim = {
  200000000: 'No',
  200000001: 'Yes'
} as const;
export type Mserp_hcmworkerentitiesmserp_primarycontactemailisim = keyof typeof Mserp_hcmworkerentitiesmserp_primarycontactemailisim;
export const Mserp_hcmworkerentitiesmserp_primarycontactemailisprivate = {
  200000000: 'No',
  200000001: 'Yes'
} as const;
export type Mserp_hcmworkerentitiesmserp_primarycontactemailisprivate = keyof typeof Mserp_hcmworkerentitiesmserp_primarycontactemailisprivate;
export const Mserp_hcmworkerentitiesmserp_primarycontactfacebookisprivate = {
  200000000: 'No',
  200000001: 'Yes'
} as const;
export type Mserp_hcmworkerentitiesmserp_primarycontactfacebookisprivate = keyof typeof Mserp_hcmworkerentitiesmserp_primarycontactfacebookisprivate;
export const Mserp_hcmworkerentitiesmserp_primarycontactfaxisprivate = {
  200000000: 'No',
  200000001: 'Yes'
} as const;
export type Mserp_hcmworkerentitiesmserp_primarycontactfaxisprivate = keyof typeof Mserp_hcmworkerentitiesmserp_primarycontactfaxisprivate;
export const Mserp_hcmworkerentitiesmserp_primarycontactlinkedinisprivate = {
  200000000: 'No',
  200000001: 'Yes'
} as const;
export type Mserp_hcmworkerentitiesmserp_primarycontactlinkedinisprivate = keyof typeof Mserp_hcmworkerentitiesmserp_primarycontactlinkedinisprivate;
export const Mserp_hcmworkerentitiesmserp_primarycontactphoneismobile = {
  200000000: 'No',
  200000001: 'Yes'
} as const;
export type Mserp_hcmworkerentitiesmserp_primarycontactphoneismobile = keyof typeof Mserp_hcmworkerentitiesmserp_primarycontactphoneismobile;
export const Mserp_hcmworkerentitiesmserp_primarycontactphoneisprivate = {
  200000000: 'No',
  200000001: 'Yes'
} as const;
export type Mserp_hcmworkerentitiesmserp_primarycontactphoneisprivate = keyof typeof Mserp_hcmworkerentitiesmserp_primarycontactphoneisprivate;
export const Mserp_hcmworkerentitiesmserp_primarycontacttwitterisprivate = {
  200000000: 'No',
  200000001: 'Yes'
} as const;
export type Mserp_hcmworkerentitiesmserp_primarycontacttwitterisprivate = keyof typeof Mserp_hcmworkerentitiesmserp_primarycontacttwitterisprivate;
export const Mserp_hcmworkerentitiesmserp_primarycontacturlisprivate = {
  200000000: 'No',
  200000001: 'Yes'
} as const;
export type Mserp_hcmworkerentitiesmserp_primarycontacturlisprivate = keyof typeof Mserp_hcmworkerentitiesmserp_primarycontacturlisprivate;
export const Mserp_hcmworkerentitiesmserp_workerstatus = {
  200000000: 'Pending',
  200000001: 'Employed',
  200000002: 'Terminated'
} as const;
export type Mserp_hcmworkerentitiesmserp_workerstatus = keyof typeof Mserp_hcmworkerentitiesmserp_workerstatus;
export const Mserp_hcmworkerentitiesmserp_workertype = {
  200000000: 'Both',
  200000001: 'Employee',
  200000002: 'Contractor'
} as const;
export type Mserp_hcmworkerentitiesmserp_workertype = keyof typeof Mserp_hcmworkerentitiesmserp_workertype;
export const Mserp_hcmworkerentitiesmserp_worksfromhome = {
  200000000: 'No',
  200000001: 'Yes'
} as const;
export type Mserp_hcmworkerentitiesmserp_worksfromhome = keyof typeof Mserp_hcmworkerentitiesmserp_worksfromhome;

export interface Mserp_hcmworkerentitiesBase {
  mserp_addressbooks?: string;
  mserp_addresscity?: string;
  mserp_addresscountryregionid?: string;
  mserp_addresscountryregionisocode?: string;
  mserp_addresscounty?: string;
  mserp_addressdistrictname?: string;
  mserp_addresslocationid?: string;
  mserp_addressnamedescription?: string;
  mserp_addresspurpose?: string;
  mserp_addressstate?: string;
  mserp_addressstreet?: string;
  mserp_addressvalidfrom?: string;
  mserp_addressvalidto?: string;
  mserp_addresszipcode?: string;
  mserp_allowrehire?: Mserp_hcmworkerentitiesmserp_allowrehire;
  mserp_anniversarydatetime?: string;
  mserp_birthdate?: string;
  mserp_citizenshipcountryregion?: string;
  mserp_deceaseddate?: string;
  mserp_disabledverificationdate?: string;
  mserp_education?: string;
  mserp_electroniclocationid?: string;
  mserp_ethnicoriginid?: string;
  mserp_expatriaterulingvalidfrom?: string;
  mserp_expatriaterulingvalidto?: string;
  mserp_fatherbirthcountryregion?: string;
  mserp_firstname?: string;
  "mserp_FK_Person_id@odata.bind"?: string;
  mserp_gender?: Mserp_hcmworkerentitiesmserp_gender;
  mserp_hcmworkerentityid: string;
  mserp_isdisabled?: Mserp_hcmworkerentitiesmserp_isdisabled;
  mserp_isdisabledveteran?: Mserp_hcmworkerentitiesmserp_isdisabledveteran;
  mserp_isexpatriaterulingapplicable?: Mserp_hcmworkerentitiesmserp_isexpatriaterulingapplicable;
  mserp_isfulltimestudent?: Mserp_hcmworkerentitiesmserp_isfulltimestudent;
  mserp_knownas?: string;
  mserp_languageid?: string;
  mserp_lastname?: string;
  mserp_lastnameprefix?: string;
  mserp_maritalstatus?: Mserp_hcmworkerentitiesmserp_maritalstatus;
  mserp_middlename?: string;
  mserp_militaryserviceenddate?: string;
  mserp_militaryservicestartdate?: string;
  mserp_motherbirthcountryregion?: string;
  mserp_namealias?: string;
  mserp_namesequencedisplayas?: string;
  mserp_nationalitycountryregion?: string;
  mserp_nativelanguageid?: string;
  mserp_numberofdependents?: number;
  mserp_officelocation?: string;
  mserp_officelocationid?: string;
  mserp_originalhiredatetime?: string;
  mserp_partynumber?: string;
  mserp_personalsuffix?: string;
  mserp_personaltitle?: string;
  mserp_personbirthcity?: string;
  mserp_personbirthcountryregion?: string;
  mserp_persondetailsvalidfrom?: string;
  mserp_persondetailsvalidto?: string;
  mserp_personnelnumber?: string;
  mserp_phoneticfirstname?: string;
  mserp_phoneticlastname?: string;
  mserp_phoneticmiddlename?: string;
  mserp_primaryaddresslocation?: number;
  mserp_primaryaddresslocation_bigint?: number;
  mserp_primarycontactemail?: string;
  mserp_primarycontactemaildescription?: string;
  mserp_primarycontactemailisim?: Mserp_hcmworkerentitiesmserp_primarycontactemailisim;
  mserp_primarycontactemailisprivate?: Mserp_hcmworkerentitiesmserp_primarycontactemailisprivate;
  mserp_primarycontactemailpurpose?: string;
  mserp_primarycontactfacebook?: string;
  mserp_primarycontactfacebookdescription?: string;
  mserp_primarycontactfacebookisprivate?: Mserp_hcmworkerentitiesmserp_primarycontactfacebookisprivate;
  mserp_primarycontactfacebookpurpose?: string;
  mserp_primarycontactfax?: string;
  mserp_primarycontactfaxdescription?: string;
  mserp_primarycontactfaxextension?: string;
  mserp_primarycontactfaxisprivate?: Mserp_hcmworkerentitiesmserp_primarycontactfaxisprivate;
  mserp_primarycontactfaxpurpose?: string;
  mserp_primarycontactlinkedin?: string;
  mserp_primarycontactlinkedindescription?: string;
  mserp_primarycontactlinkedinisprivate?: Mserp_hcmworkerentitiesmserp_primarycontactlinkedinisprivate;
  mserp_primarycontactlinkedinpurpose?: string;
  mserp_primarycontactphone?: string;
  mserp_primarycontactphonedescription?: string;
  mserp_primarycontactphoneextension?: string;
  mserp_primarycontactphoneismobile?: Mserp_hcmworkerentitiesmserp_primarycontactphoneismobile;
  mserp_primarycontactphoneisprivate?: Mserp_hcmworkerentitiesmserp_primarycontactphoneisprivate;
  mserp_primarycontactphonepurpose?: string;
  mserp_primarycontacttwitter?: string;
  mserp_primarycontacttwitterdescription?: string;
  mserp_primarycontacttwitterisprivate?: Mserp_hcmworkerentitiesmserp_primarycontacttwitterisprivate;
  mserp_primarycontacttwitterpurpose?: string;
  mserp_primarycontacturl?: string;
  mserp_primarycontacturldescription?: string;
  mserp_primarycontacturlisprivate?: Mserp_hcmworkerentitiesmserp_primarycontacturlisprivate;
  mserp_primarycontacturlpurpose?: string;
  mserp_professionalsuffix?: string;
  mserp_professionaltitle?: string;
  mserp_senioritydate?: string;
  mserp_summaryvalidfrom?: string;
  mserp_summaryvalidto?: string;
  mserp_titleid?: string;
  mserp_veteranstatusid?: string;
  mserp_worksfromhome?: Mserp_hcmworkerentitiesmserp_worksfromhome;
}

export interface Mserp_hcmworkerentities extends Mserp_hcmworkerentitiesBase {
  mserp_allowrehirename?: string;
  mserp_fk_person_idname?: string;
  mserp_gendername?: string;
  mserp_identityemail?: string;
  mserp_identityprovider?: string;
  mserp_isdisabledname?: string;
  mserp_isdisabledveteranname?: string;
  mserp_isexpatriaterulingapplicablename?: string;
  mserp_isfulltimestudentname?: string;
  mserp_maritalstatusname?: string;
  mserp_name?: string;
  mserp_objectid?: string;
  mserp_partytype?: string;
  mserp_personuservalidfrom?: string;
  mserp_personuservalidto?: string;
  mserp_primarycontactemailisimname?: string;
  mserp_primarycontactemailisprivatename?: string;
  mserp_primarycontactfacebookisprivatename?: string;
  mserp_primarycontactfaxisprivatename?: string;
  mserp_primarycontactlinkedinisprivatename?: string;
  mserp_primarycontactphoneismobilename?: string;
  mserp_primarycontactphoneisprivatename?: string;
  mserp_primarycontacttwitterisprivatename?: string;
  mserp_primarycontacturlisprivatename?: string;
  mserp_user?: string;
  mserp_workerstatus?: Mserp_hcmworkerentitiesmserp_workerstatus;
  mserp_workerstatusname?: string;
  mserp_workertype?: Mserp_hcmworkerentitiesmserp_workertype;
  mserp_workertypename?: string;
  mserp_worksfromhomename?: string;
  mserp_fk_person_id?: object;
  _mserp_fk_person_id_value?: string;
}